% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cinhaz.R
\name{cinhaz}
\alias{cinhaz}
\title{Conduct partial and total R2-based sensitivity analysis}
\usage{
cinhaz(kd, ky, data, outcome, treatment, bnch_reg, other_reg, alpha)
}
\arguments{
\item{kd}{relative strength of the confounder in explaining variation in treatment as compared to benchmark covariate(s)}

\item{ky}{relative strength of confounder in explaining variation in outcome as compared to benchmerk covariate(s)}

\item{data}{data frame}

\item{outcome}{outcome variable}

\item{treatment}{treatment variable}

\item{bnch_reg}{benchmark covariate(s)}

\item{other_reg}{other covariates in the model (other than treatment and benchmark covariates)}

\item{alpha}{significance level for hypothesis test (H0: true effect = 0)}
}
\value{
A data frame with results
}
\description{
Conduct partial and total R2-based sensitivity analysis
}
\examples{
## Load library
library(sensemakr)
## Conduct analysis
cinhaz(kd=1,ky=1,data=darfur,outcome = "peacefactor",
treatment = "directlyharmed", bnch_reg = "female",
other_reg = c("village","age","farmer_dar","herder_dar","pastvoted","hhsize_darfur"),
alpha=0.05)
}
