% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cplotbias.R
\name{cplotbias}
\alias{cplotbias}
\title{Create contour plot of bias}
\usage{
cplotbias(data)
}
\arguments{
\item{data}{A data frame that is the output from the "ovbias" function.}
}
\value{
A plot object created with ggplot
}
\description{
Create contour plot of bias
}
