% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delfplot.R
\name{delfplot}
\alias{delfplot}
\title{Plot graph of function delta=f(Rmax)}
\usage{
delfplot(parameters)
}
\arguments{
\item{parameters}{A vector of parameters that is generated after estimating the short, intermediate and auxiliary regressions.}
}
\value{
A plot object created with ggplot
}
\description{
Plot graph of function delta=f(Rmax)
}
