% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_border.R
\name{expand_border}
\alias{expand_border}
\title{Extend border of bounded box by +/- e}
\usage{
expand_border(parameters, deltalow, deltahigh, Rlow, Rhigh, e)
}
\arguments{
\item{parameters}{A vector of parameters (real numbers) that is generated by estimating the short, intermediate and auxiliary regressions.}

\item{deltalow}{The lower limit of delta.}

\item{deltahigh}{The upper limit of delta.}

\item{Rlow}{The lower limit of Rmax.}

\item{Rhigh}{The upper limit of Rmax.}

\item{e}{The step size.}
}
\value{
Data frame.
}
\description{
Extend border of bounded box by +/- e
}
