% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_border.R
\name{get_border}
\alias{get_border}
\title{Identify all border points in a region}
\usage{
get_border(region, e)
}
\arguments{
\item{region}{A data frame containing the x and y coordinates of the region.}

\item{e}{The step size of the grid in the x and y directions.}
}
\value{
A data frame containing the x and y coordinates of the border points of the region.
}
\description{
Identify all border points in a region
}
