% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycubic.R
\name{mycubic}
\alias{mycubic}
\title{Compute roots of the cubic equation}
\usage{
mycubic(parameters, mydelta, Rmax)
}
\arguments{
\item{parameters}{A vector of parameters (real numbers) that is generated by estimating the short, intermediate and auxiliary regressions.}

\item{mydelta}{Value of delta (real number).}

\item{Rmax}{Value of Rmax (real number).}
}
\value{
A vector containing the three roots of the cubic equation defined by the parameters, delta and Rmax.
}
\description{
Compute roots of the cubic equation
}
