% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydisc.R
\name{mydisc}
\alias{mydisc}
\title{Evaluates discriminant of the cubic equation}
\usage{
mydisc(parameters, mydelta, Rmax)
}
\arguments{
\item{parameters}{A vector of parameters (real numbers) that is generated by estimating the short, intermediate and auxiliary regressions.}

\item{mydelta}{The value of delta (real number).}

\item{Rmax}{The value of Rmax (real number)}
}
\value{
Returns a value of 0 or 1; 0 (if discriminant is positive) and 1 (if discriminant is nonpositive)
}
\description{
Evaluates discriminant of the cubic equation
}
