% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NLSY_IQ}
\alias{NLSY_IQ}
\title{NLSY IQ.}
\format{
A data frame with twelve variables:
\describe{
\item{\code{iq_std}}{standardized IQ score, PIAT score}
\item{\code{BF_months}}{months of breast feeding}
\item{\code{mom_drink_preg_all}}{did mother drink at all during pregnancy}
\item{\code{lbw_preterm}}{low birth weight + preterm}
\item{\code{age}}{age of child}
\item{\code{female}}{child female}
\item{\code{black}}{mother black}
\item{\code{motherAge}}{age of mother}
\item{\code{motherEDU}}{years of schooling of mother}
\item{\code{mom_married}}{is the mother married?}
\item{\code{income}}{annual income of mother}
\item{\code{sex}}{child sex}
\item{\code{race}}{race of mother}
}

Source: \url{https://drive.google.com/file/d/1O1W9dP8F3B1DnAZGBegpoqCfysUrn7Uc/view?usp=sharing}
}
\usage{
NLSY_IQ
}
\description{
NLSY data to analyse the effect of maternal behaviour on
children's IQ score. Natality detail files are from 2001 and 2002.
Data is from the NLSY Children and Young Adults panel.
}
\keyword{datasets}
