% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovbias.R
\name{ovbias}
\alias{ovbias}
\title{Compute bias adjusted treatment effect taking parameter vector as input.}
\usage{
ovbias(parameters, deltalow, deltahigh, Rhigh, e)
}
\arguments{
\item{parameters}{A vector of parameters (real numbers) that is generated by estimating the short, intermediate and auxiliary regressions.}

\item{deltalow}{The lower limit of delta.}

\item{deltahigh}{The upper limit of delta.}

\item{Rhigh}{The upper limit of Rmax.}

\item{e}{The step size.}
}
\value{
List with three elements:

\item{Data}{Data frame containing the bias and bias-adjusted treatment effect for each point on the grid}
\item{bias_Distribution}{Quantiles (2.5,5.0,50,95,97.5) of the empirical distribution of bias}
\item{bstar_Distribution}{Quantiles (2.5,5.0,50,95,97.5) of the empirical distribution of the bias-adjusted treatment effect}
}
\description{
Compute bias adjusted treatment effect taking parameter vector as input.
}
