% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partocoef.R
\name{partocoef}
\alias{partocoef}
\title{Returns coefficients of the cubic equation}
\usage{
partocoef(parameters, mydelta, Rmax)
}
\arguments{
\item{parameters}{A vector of parameters (real numbers) that is generated by estimating the short, intermediate and auxiliary regressions.}

\item{mydelta}{The value of delta (real number)}

\item{Rmax}{The value of Rmax (real number)}
}
\value{
A data frame with the coefficients of the cubic equation.
}
\description{
Returns coefficients of the cubic equation
}
