% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_root.R
\name{selectroot}
\alias{selectroot}
\title{Select root of the cubic based on the root of a nearest point}
\usage{
selectroot(parameters, mydelta, Rmax, closest_bias)
}
\arguments{
\item{parameters}{A vector of parameters (real numbers) that is generated by estimating the short, intermediate and auxiliary regressions.}

\item{mydelta}{The value of delta (real number).}

\item{Rmax}{The value of Rmax (real number).}

\item{closest_bias}{The value of bias at the nearest point.}
}
\value{
Data frame
}
\description{
Select root of the cubic based on the root of a nearest point
}
