% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_nurr.R
\name{split_nurr}
\alias{split_nurr}
\title{Split a region into two parts}
\usage{
split_nurr(region1, region2, epsilon, parameters, e)
}
\arguments{
\item{region1}{Data frame with coordinates for region 1}

\item{region2}{Data frame with coordinates for region 2}

\item{epsilon}{Closest distance}

\item{parameters}{A vector of parameters (real numbers) that is generated by estimating the short, intermediate and auxiliary regressions.}

\item{e}{The step size of the grid in the x and y directions.}
}
\value{
List, where first element is region within epsilon distance of region 1 and second element which is region which is not within epsilon distance of region 1.
}
\description{
Split a region into two parts
}
