% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplotbate.R
\name{dplotbate}
\alias{dplotbate}
\title{Histogram of bias adjusted treatment effect}
\usage{
dplotbate(data)
}
\arguments{
\item{data}{A data frame that is the output from the "ovbias" function.}
}
\value{
A plot object created with ggplot
}
\description{
Histogram of bias adjusted treatment effect
}
