% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urrplot.R
\name{urrplot}
\alias{urrplot}
\title{Region plot to demarcate URR and NURR for the bounded box}
\usage{
urrplot(parameters, deltalow, deltahigh, Rlow, Rhigh, e)
}
\arguments{
\item{parameters}{A vector of parameters (real numbers) that is generated by estimating the short, intermediate and auxiliary regressions.}

\item{deltalow}{The lower limit for delta.}

\item{deltahigh}{The upper limit for delta.}

\item{Rlow}{The lower limit for Rmax.}

\item{Rhigh}{The upper limit for Rmax.}

\item{e}{The step size of the grid in the x and y directions.}
}
\value{
A plot object created by ggplot
}
\description{
Region plot to demarcate URR and NURR for the bounded box
}
